move mock data to mock API call
