+++ my-app/src/screens/BatchSelectionScreen.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from 'react';
 import { ScrollView, View } from 'react-native';
 import HorizontalSelector from '../components/HorizontalSelector';
 import { SearchBar } from '../components/SearchBar';
-import { getBatchYears } from '../remote/CaliberBatchAPI';
+import { getBatchYears, getBatchesByYear } from '../remote/CaliberBatchAPI';
 import Batch from '../models/batch';
 import { styles1 } from '../styles/style1';
 import BatchList from '../components/BatchList';
@@ -12,71 +12,10 @@ type Props = {
 
 }
 
-const DATA:Array<Batch> = [
-  {
-    batchId: '123123',
-    batchTitle: '05132021 Cloud Native Matt',
-    trainers: [
-      {
-        username: 'Matt',
-        role: 'Trainer',
-      },
-    ],
-    startDate: new Date().toISOString(),
-    endDate: new Date().toISOString(),
-    associates: [
-
-    ],
-    notes: [
-
-    ],
-  },
-  {
-    batchId: 'tryhtry',
-    batchTitle: '05132021 DevOps Tim',
-    trainers: [
-      {
-        username: 'Tim',
-        role: 'Trainer',
-      },
-    ],
-    startDate: new Date().toISOString(),
-    endDate: new Date().toISOString(),
-    associates: [
-
-    ],
-    notes: [
-
-    ],
-  },
-];
-
-for (let i = 0; i < 20; i += 1) {
-  const newBatch: Batch = {
-    batchId: `${i}`,
-    batchTitle: '05132020 DevOps Tim',
-    trainers: [
-      {
-        username: `${i % 2 === 0 ? 'Matt' : 'Tim'}`,
-        role: 'Trainer',
-      },
-    ],
-    startDate: new Date().toISOString(),
-    endDate: new Date().toISOString(),
-    associates: [
-
-    ],
-    notes: [
-
-    ],
-  };
-  DATA.push(newBatch);
-}
-
 const BatchSelectionScreen: React.FC<Props> = (): JSX.Element => {
   const navigation = useNavigation();
   const [years, setYears] = useState<string[]>([]);
-  const [selectedYear, setSelectedYear] = useState<string>();
+  const [selectedYear, setSelectedYear] = useState<string>(new Date().getFullYear().toString());
   const [batchList, setBatchList] = useState<Batch[]>([]);
   const [searchResults, setSearchResults] = useState<Batch[]>();
 
@@ -91,7 +30,13 @@ const BatchSelectionScreen: React.FC<Props> = (): JSX.Element => {
 
   useEffect(() => {
     // set batch list according to selected year
-    setBatchList(DATA);
+    const getBatchList = async (): Promise<void> => {
+      if (selectedYear) {
+        setBatchList(await getBatchesByYear(selectedYear));
+      }
+    };
+    getBatchList();
+
     setSearchResults(undefined);
   }, [selectedYear]);
